+++ src/Exceptions/FileCannotBeAdded/FileUnacceptableForCollection.php
@@ -3,7 +3,7 @@
 namespace Spatie\Medialibrary\Exceptions\FileCannotBeAdded;
 
 use Spatie\Medialibrary\Exceptions\FileCannotBeAdded\FileCannotBeAdded;
-use Spatie\Medialibrary\File;
+use Spatie\Medialibrary\HasMedia\File;
 use Spatie\Medialibrary\HasMedia\HasMedia;
 use Spatie\Medialibrary\HasMedia\MediaCollection;