+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `laravel-medialibrary` will be documented in this file
 
+## 7.19.1 - 2020-03-04
+
+- Update S3 url generator to use media disk #1755
+
 ## 7.19.0 - 2020-03-03
 
 - add support for Laravel 7